requirements.txt
Flask
requests
transformers
torch
python-dotenv
docx2txt
pypdf2
sentencepiece